+++ src/parsers/parseCronRange.ts
@@ -1,4 +1,4 @@
-import { isUndefined } from "@vangware/utils";
+import { isUndefined } from "@vangware/predicates";
 import type { CronRange } from "../types/CronRange.js";
 import type { CronValueParser } from "../types/CronValueParser.js";
 import { isCronRange } from "../validations/isCronRange.js";
@@ -9,13 +9,10 @@ import { isCronRange } from "../validations/isCronRange.js";
  * @category Parser
  * @param parser `CronValueParser` for `CronRange`.
  * @returns Curried function with `parser` in context.
+ * @example
  */
 export const parseCronRange =
 	<Value>(parser: CronValueParser<Value>) =>
-	/**
-	 * @param source `CronRange` to be parsed.
-	 * @returns A string or `undefined` if invalid.
-	 */
 	(source: CronRange<Value>) => {
 		const valid = isCronRange(source);
 		const from = valid ? parser(source.from) : undefined;