+++ src/validations/isCronStartOrBlank.ts
@@ -1,4 +1,4 @@
-import { isString } from "@vangware/utils";
+import { isString } from "@vangware/predicates";
 import { CRON_START_OR_BLANK } from "../constants.js";
 import type { CronStartOrBlank } from "../types/CronStartOrBlank.js";
 
@@ -7,6 +7,7 @@ import type { CronStartOrBlank } from "../types/CronStartOrBlank.js";
  *
  * @category Validation
  * @param value Value to check.
+ * @example
  */
 export const isCronStartOrBlank = (value: unknown): value is CronStartOrBlank =>
 	isString(value) && value === CRON_START_OR_BLANK;