+++ tests/parsers/parseCronDayOfWeek.test.ts
@@ -56,9 +56,7 @@ export default [
 		wanted: "MON-SAT",
 	},
 	{
-		given:
-			// eslint-disable-next-line max-len
-			"on every Sunday, Monday and Tuesday, and every day between Thursday and Saturday",
+		given: "on every Sunday, Monday and Tuesday, and every day between Thursday and Saturday",
 		must: "return CronList",
 		received: parseCronDayOfWeek([1, 2, 3, { from: 5, to: 7 }]),
 		wanted: "1,2,3,5-7",