+++ tests/parsers/parseStringDayOfMonth.test.ts
@@ -50,9 +50,7 @@ export default [
 		wanted: { from: 5, to: 10 },
 	},
 	{
-		given:
-			// eslint-disable-next-line max-len
-			"on the 1st, 2nd, 3rd and 4th day and every day between 5th and 10th",
+		given: "on the 1st, 2nd, 3rd and 4th day and every day between 5th and 10th",
 		must: "return CronList",
 		received: parseStringDayOfMonth("1,2,3,4,5-10"),
 		wanted: [1, 2, 3, 4, { from: 5, to: 10 }],